Remove std::pow from reduction. (#2234)
This is a functional simplification that removes the std::pow from
reduction. The resulting reduction values are within 1% of master.

This is a simplification because i believe an fp addition and
multiplication is much faster than a call to std::pow() which is
historically slow and performance varies widely on different
architectures.

STC
LLR: 2.95 (-2.94,2.94) [-3.00,1.00]
Total: 23471 W: 5245 L: 5127 D: 13099
http://tests.stockfishchess.org/tests/view/5d27ac1b0ebc5925cf0d476b

LTC
LLR: 2.95 (-2.94,2.94) [-3.00,1.00]
Total: 51533 W: 8736 L: 8665 D: 34132
http://tests.stockfishchess.org/tests/view/5d27b74e0ebc5925cf0d493c

Bench 3765158
master (#2234)
@protonspring
@mcostalba
protonspring authored